Create generic FileStoreHandler
