+++ tests/Services/FileStoreHandler.php
@@ -1,0 +1,47 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Tests\Services;
+
+use Symfony\Component\Finder\Finder;
+
+class FileStoreHandler
+{
+    public function __construct(
+        private BasilFixtureHandler $basilFixtureHandler,
+        private string $path
+    ) {
+    }
+
+    public function clear(): void
+    {
+        $finder = new Finder();
+        $finder->files()->in($this->path);
+
+        foreach ($finder as $file) {
+            $path = $file->getPathname();
+
+            if (file_exists($path)) {
+                unlink($path);
+            }
+        }
+    }
+
+    public function copyFixture(string $relativePath): string
+    {
+        $origin = $this->basilFixtureHandler->getPath($relativePath);
+        $destination = $this->path . '/' . $relativePath;
+
+        if (!file_exists($destination)) {
+            $directory = dirname($destination);
+            if (!file_exists($directory)) {
+                mkdir($directory, 0777, true);
+            }
+
+            copy($origin, $destination);
+        }
+
+        return $destination;
+    }
+}
